Внес мелкие исправления по неймингу
