+++ src/utils/common.js
@@ -1,3 +1,5 @@
+import dayjs from 'dayjs';
+
 // Функция из интернета по генерации случайного числа из диапазона
 // Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
 const getRandomInteger = (a = 0, b = 1) => {
@@ -19,4 +21,16 @@ const shuffle = (array) => {
 
 const isEscEvent = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
 
-export {getRandomInteger, shuffle, isEscEvent};
+const generateDate = () => {
+  const maxDaysGap = 7;
+  const daysGap = getRandomInteger(-maxDaysGap, maxDaysGap);
+
+  return dayjs().add(daysGap, 'day').toDate();
+};
+
+const showFullDate = (date) => dayjs(date).format('D MMMM YYYY');
+
+//Функция для добавления окончания "s", если существительное во множественном числе
+const isMultiple = (data) => data.length > 1 ? 's' : '';
+
+export {getRandomInteger, shuffle, isEscEvent, generateDate, showFullDate, isMultiple};
