+++ src/view/films-list.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract.js';
+import AbstractView from './abstract-view.js';
 
 const createFilmsListTemplate = () => (
   `<section class="films">
