+++ src/view/footer-statistics.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract.js';
+import AbstractView from './abstract-view.js';
 
 const createFooterStatisticsTemplate = (films) => `<p>${films.length} movies inside</p>`;
