+++ src/view/no-films.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract.js';
+import AbstractView from './abstract-view.js';
 
 const createNoFilmsTemplate = () => (
   `<section class="films">
